requirements.txt
fastapi
uvicorn[standard]
langchain
openai
chromadb
tiktoken
python-multipart
pydantic
python-dotenv
requests
